1,全部图片解析完成;2,增加大炮相关

+++ game/game-api/src/main/java/net/keshen/base/graphics/Matrix.java
@@ -12,7 +12,7 @@ public interface Matrix {
 	 * @param x
 	 * @param y
 	 */
-	public void setTranslate(int x,int y);
+	public void setTranslate(float x,float y);
 	
 	/**
 	 * 重置矩阵
